Refactored and successfully removed ClassicTests, AsyncTests, fixture.ClassicTests, fixture.AsyncTests.
