+++ scalatest/src/main/scala/org/scalatest/AsyncFunSuiteLike.scala
@@ -18,7 +18,7 @@ package org.scalatest
 import scala.concurrent.{ExecutionContext, Future}
 
 //SCALATESTJS-ONLY @scala.scalajs.js.annotation.JSExportDescendentClasses(ignoreInvalidDescendants = true)
-trait AsyncFunSuiteLike extends FunSuiteRegistration with AsyncTests with OneInstancePerTest { thisSuite =>
+trait AsyncFunSuiteLike extends FunSuiteOf[Future[Unit]] with AsyncFixtures with OneInstancePerTest { thisSuite =>
 
   implicit def executionContext: ExecutionContext
