+++ scalatest/src/main/scala/org/scalatest/AsyncPropSpecLike.scala
@@ -18,7 +18,7 @@ package org.scalatest
 import scala.concurrent.{ExecutionContext, Future}
 
 //SCALATESTJS-ONLY @scala.scalajs.js.annotation.JSExportDescendentClasses(ignoreInvalidDescendants = true)
-trait AsyncPropSpecLike extends PropSpecRegistration with AsyncTests with OneInstancePerTest { thisSuite =>
+trait AsyncPropSpecLike extends PropSpecOf[Future[Unit]] with AsyncFixtures with OneInstancePerTest { thisSuite =>
 
   implicit def executionContext: ExecutionContext
