+++ scalatest/src/main/scala/org/scalatest/AsyncWordSpecLike.scala
@@ -18,7 +18,7 @@ package org.scalatest
 import scala.concurrent.{ExecutionContext, Future}
 
 //SCALATESTJS-ONLY @scala.scalajs.js.annotation.JSExportDescendentClasses(ignoreInvalidDescendants = true)
-trait AsyncWordSpecLike extends WordSpecRegistration with AsyncTests with OneInstancePerTest { thisSuite =>
+trait AsyncWordSpecLike extends WordSpecOf[Future[Unit]] with AsyncFixtures with OneInstancePerTest { thisSuite =>
 
   implicit def executionContext: ExecutionContext
