+++ scalatest/src/main/scala/org/scalatest/FlatSpecLike.scala
@@ -16,4 +16,4 @@
 package org.scalatest
 
 //SCALATESTJS-ONLY @scala.scalajs.js.annotation.JSExportDescendentClasses(ignoreInvalidDescendants = true)
-trait FlatSpecLike extends FlatSpecRegistration with ClassicTests
+trait FlatSpecLike extends FlatSpecOf[Unit]
