+++ scalatest/src/main/scala/org/scalatest/FunSpecOf.scala
@@ -13,50 +13,47 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest.fixture
+package org.scalatest
 
-import org.scalatest._
 import scala.collection.immutable.ListSet
-import org.scalatest.exceptions.StackDepthExceptionHelper.getStackDepthFun
-import java.util.concurrent.atomic.AtomicReference
-import java.util.ConcurrentModificationException
-import org.scalatest.events._
-import org.scalatest.Suite.anExceptionThatShouldCauseAnAbort
-import org.scalatest.Suite.autoTagClassAnnotations
 import words.BehaveWord
-import org.scalatest.exceptions.TestRegistrationClosedException
+import Suite.autoTagClassAnnotations
 
 /**
- * Implementation trait for class <code>fixture.FunSpec</code>, which is
- * a sister class to <code>org.scalatest.FunSpec</code> that can pass a
- * fixture object into its tests.
+ * Implementation trait for class <code>FunSpec</code>, which
+ * facilitates a &ldquo;behavior-driven&rdquo; style of development (BDD),
+ * in which tests are combined with text that specifies the behavior the tests
+ * verify.
  *
  * <p>
- * <a href="FunSpec.html"><code>fixture.FunSpec</code></a> is a class,
- * not a trait, to minimize compile time given there is a slight compiler
- * overhead to mixing in traits compared to extending classes. If you need
- * to mix the behavior of <code>fixture.FunSpec</code> into some other
- * class, you can use this trait instead, because class
- * <code>fixture.FunSpec</code> does nothing more than extend this trait and add a nice <code>toString</code> implementation.
+ * <a href="FunSpec.html"><code>FunSpec</code></a> is a class, not a trait,
+ * to minimize compile time given there is a slight compiler overhead to
+ * mixing in traits compared to extending classes. If you need to mix the
+ * behavior of <code>FunSpec</code> into some other class, you can use this
+ * trait instead, because class <code>FunSpec</code> does nothing more than
+ * extend this trait and add a nice <code>toString</code> implementation.
  * </p>
  *
  * <p>
  * See the documentation of the class for a <a href="FunSpec.html">detailed
- * overview of <code>fixture.FunSpec</code></a>.
+ * overview of <code>FunSpec</code></a>.
  * </p>
  *
  * @author Bill Venners
  */
 @Finders(Array("org.scalatest.finders.FunSpecFinder"))
-trait FunSpecRegistration extends Suite with TestRegistration with Informing with Notifying with Alerting with Documenting { thisSuite =>
+trait FunSpecOf[R] extends Suite with TestRegistration with Informing with Notifying with Alerting with Documenting { thisSuite =>
 
-  private final val engine = new FixtureEngine[FixtureParam](Resources.concurrentFixtureSpecMod, "FixtureFunSpec")
+  type Registration = R
 
-  protected[scalatest] def getEngine: FixtureEngine[FixtureParam] = engine
+  private final val engine = new Engine(Resources.concurrentSpecMod, "FunSpec")
+
+  protected[scalatest] def getEngine: Engine = engine
 
   import engine._
 
-  private[scalatest] val sourceFileName = "FunSpecRegistration.scala"
+  // TODO: Probably make this private final val sourceFileName in a singleton object so it gets compiled in rather than carried around in each instance
+  private[scalatest] val sourceFileName = "FunSpecOf.scala"
 
   /**
    * Returns an <code>Informer</code> that during test execution will forward strings passed to its
@@ -100,7 +97,7 @@ trait FunSpecRegistration extends Suite with TestRegistration with Informing wit
    */
   protected def markup: Documenter = atomicDocumenter.get
 
-  final def registerTest(testText: String, testTags: Tag*)(testFun: FixtureParam => Registration) {
+  final def registerTest(testText: String, testTags: Tag*)(testFun: => Registration) {
     // SKIP-SCALATESTJS-START
     val stackDepthAdjustment = -2
     // SKIP-SCALATESTJS-END
@@ -108,12 +105,12 @@ trait FunSpecRegistration extends Suite with TestRegistration with Informing wit
     engine.registerTest(testText, transformToOutcome(testFun), Resources.testCannotBeNestedInsideAnotherTest, sourceFileName, "registerTest", 5, stackDepthAdjustment, None, None, None, testTags: _*)
   }
 
-  final def registerIgnoredTest(testText: String, testTags: Tag*)(testFun: FixtureParam => Registration) {
+  final def registerIgnoredTest(testText: String, testTags: Tag*)(testFun: => Registration) {
     // SKIP-SCALATESTJS-START
-    val stackDepthAdjustment = 0
+    val stackDepthAdjustment = -2
     // SKIP-SCALATESTJS-END
-    //SCALATESTJS-ONLY val stackDepthAdjustment = -2
-    engine.registerIgnoredTest(testText, transformToOutcome(testFun), Resources.testCannotBeNestedInsideAnotherTest, sourceFileName, "registerIgnoredTest", 1, stackDepthAdjustment, None, testTags: _*)
+    //SCALATESTJS-ONLY val stackDepthAdjustment = -5
+    engine.registerIgnoredTest(testText, transformToOutcome(testFun), Resources.testCannotBeNestedInsideAnotherTest, sourceFileName, "registerIgnoredTest", 4, stackDepthAdjustment, None, testTags: _*)
   }
 
   /**
@@ -121,24 +118,28 @@ trait FunSpecRegistration extends Suite with TestRegistration with Informing wit
    * supports test (and shared test) registration in <code>FunSpec</code>s.
    *
    * <p>
-   * This class supports syntax such as the following:
+   * This class supports syntax such as the following test registration:
    * </p>
    *
-   * <pre class="stHighlight">
+   * <pre class="stExamples">
    * it("should be empty")
    * ^
    * </pre>
    *
-   * <pre class="stHighlight">
+   * <p>
+   * and the following shared test registration:
+   * </p>
+   *
+   * <pre class="stExamples">
    * it should behave like nonFullStack(stackWithOneItem)
    * ^
    * </pre>
    *
    * <p>
-   * For more information and examples, see the <a href="../FunSpec.html">main documentation for <code>FunSpec</code></a>.
+   * For more information and examples, see the <a href="FunSpec.html">main documentation for <code>FunSpec</code></a>.
    * </p>
    */
-  protected final class ItWord {
+  protected class ItWord {
 
     /**
      * Register a test with the given spec text, optional tags, and test function value that takes no arguments.
@@ -158,13 +159,13 @@ trait FunSpecRegistration extends Suite with TestRegistration with Informing wit
      * @throws TestRegistrationClosedException if invoked after <code>run</code> has been invoked on this suite
      * @throws NullArgumentException if <code>specText</code> or any passed test tag is <code>null</code>
      */
-    def apply(specText: String, testTags: Tag*)(testFun: FixtureParam => Registration) {
+    def apply(specText: String, testTags: Tag*)(testFun: => Registration) {
       // SKIP-SCALATESTJS-START
       val stackDepth = 3
       val stackDepthAdjustment = -2
       // SKIP-SCALATESTJS-END
       //SCALATESTJS-ONLY val stackDepth = 5
-      //SCALATESTJS-ONLY val stackDepthAdjustment = -5
+      //SCALATESTJS-ONLY val stackDepthAdjustment = -4
       engine.registerTest(specText, transformToOutcome(testFun), Resources.itCannotAppearInsideAnotherItOrThey, sourceFileName, "apply", stackDepth, stackDepthAdjustment, None, None, None, testTags: _*)
     }
 
@@ -175,17 +176,15 @@ trait FunSpecRegistration extends Suite with TestRegistration with Informing wit
      * This method supports syntax such as the following:
      * </p>
      *
-     * <pre class="stHighlight">
+     * <pre class="stExamples">
      * it should behave like nonFullStack(stackWithOneItem)
      *    ^
      * </pre>
      *
      * <p>
-     * For examples of shared tests, see the <a href="../Spec.html#SharedTests">Shared tests section</a>
+     * For examples of shared tests, see the <a href="FunSpec.html#sharedTests">Shared tests section</a>
      * in the main documentation for trait <code>FunSpec</code>.
      * </p>
-     *
-     * @param behaveWord the <code>BehaveWord</code>
      */
     def should(behaveWord: BehaveWord) = behaveWord
 
@@ -196,17 +195,15 @@ trait FunSpecRegistration extends Suite with TestRegistration with Informing wit
      * This method supports syntax such as the following:
      * </p>
      *
-     * <pre class="stHighlight">
+     * <pre class="stExamples">
      * it must behave like nonFullStack(stackWithOneItem)
      *    ^
      * </pre>
      *
      * <p>
-     * For examples of shared tests, see the <a href="../Spec.html#SharedTests">Shared tests section</a>
+     * For examples of shared tests, see the <a href="FunSpec.html#sharedTests">Shared tests section</a>
      * in the main documentation for trait <code>FunSpec</code>.
      * </p>
-     *
-     * @param behaveWord the <code>BehaveWord</code>
      */
     def must(behaveWord: BehaveWord) = behaveWord
   }
@@ -218,19 +215,18 @@ trait FunSpecRegistration extends Suite with TestRegistration with Informing wit
    * This field supports syntax such as the following:
    * </p>
    *
-   * <pre class="stHighlight">
+   * <pre class="stExamples">
    * it("should be empty")
    * ^
    * </pre>
    *
-   * <pre class="stHighlight">
+   * <pre> class="stExamples"
    * it should behave like nonFullStack(stackWithOneItem)
    * ^
    * </pre>
    *
    * <p>
-   * For more information and examples of the use of the <code>it</code> field, see
-   * the <a href="../FunSpec.html">main documentation for <code>FunSpec</code></a>.
+   * For more information and examples of the use of the <code>it</code> field, see the main documentation for this trait.
    * </p>
    */
   protected val it = new ItWord
@@ -240,24 +236,28 @@ trait FunSpecRegistration extends Suite with TestRegistration with Informing wit
    * supports test (and shared test) registration in <code>FunSpec</code>s.
    *
    * <p>
-   * This class supports syntax such as the following:
+   * This class supports syntax such as the following test registration:
    * </p>
    *
-   * <pre class="stHighlight">
+   * <pre class="stExamples">
    * they("should be empty")
    * ^
    * </pre>
    *
-   * <pre class="stHighlight">
+   * <p>
+   * and the following shared test registration:
+   * </p>
+   *
+   * <pre class="stExamples">
    * they should behave like nonFullStack(stackWithOneItem)
    * ^
    * </pre>
    *
    * <p>
-   * For more information and examples, see the <a href="../FunSpec.html">main documentation for <code>FunSpec</code></a>.
+   * For more information and examples, see the <a href="FunSpec.html">main documentation for <code>FunSpec</code></a>.
    * </p>
    */
-  protected final class TheyWord {
+  protected class TheyWord {
 
     /**
      * Register a test with the given spec text, optional tags, and test function value that takes no arguments.
@@ -277,12 +277,8 @@ trait FunSpecRegistration extends Suite with TestRegistration with Informing wit
      * @throws TestRegistrationClosedException if invoked after <code>run</code> has been invoked on this suite
      * @throws NullArgumentException if <code>specText</code> or any passed test tag is <code>null</code>
      */
-    def apply(specText: String, testTags: Tag*)(testFun: FixtureParam => Registration) {
-      // SKIP-SCALATESTJS-START
-      val stackDepthAdjustment = -2
-      // SKIP-SCALATESTJS-END
-      //SCALATESTJS-ONLY val stackDepthAdjustment = -3
-      engine.registerTest(specText, transformToOutcome(testFun), Resources.theyCannotAppearInsideAnotherItOrThey, sourceFileName, "apply", 3, stackDepthAdjustment, None, None, None, testTags: _*)
+    def apply(specText: String, testTags: Tag*)(testFun: => Registration) {
+      engine.registerTest(specText, transformToOutcome(testFun), Resources.theyCannotAppearInsideAnotherItOrThey, sourceFileName, "apply", 3, -2, None, None, None, testTags: _*)
     }
 
     /**
@@ -292,17 +288,15 @@ trait FunSpecRegistration extends Suite with TestRegistration with Informing wit
      * This method supports syntax such as the following:
      * </p>
      *
-     * <pre class="stHighlight">
+     * <pre class="stExamples">
      * they should behave like nonFullStack(stackWithOneItem)
      *      ^
      * </pre>
      *
      * <p>
-     * For examples of shared tests, see the <a href="../Spec.html#SharedTests">Shared tests section</a>
+     * For examples of shared tests, see the <a href="FunSpec.html#sharedTests">Shared tests section</a>
      * in the main documentation for trait <code>FunSpec</code>.
      * </p>
-     *
-     * @param behaveWord the <code>BehaveWord</code>
      */
     def should(behaveWord: BehaveWord) = behaveWord
 
@@ -313,17 +307,15 @@ trait FunSpecRegistration extends Suite with TestRegistration with Informing wit
      * This method supports syntax such as the following:
      * </p>
      *
-     * <pre class="stHighlight">
+     * <pre class="stExamples">
      * they must behave like nonFullStack(stackWithOneItem)
      *      ^
      * </pre>
      *
      * <p>
-     * For examples of shared tests, see the <a href="../Spec.html#SharedTests">Shared tests section</a>
+     * For examples of shared tests, see the <a href="FunSpec.html#sharedTests">Shared tests section</a>
      * in the main documentation for trait <code>FunSpec</code>.
      * </p>
-     *
-     * @param behaveWord the <code>BehaveWord</code>
      */
     def must(behaveWord: BehaveWord) = behaveWord
   }
@@ -335,19 +327,18 @@ trait FunSpecRegistration extends Suite with TestRegistration with Informing wit
    * This field supports syntax such as the following:
    * </p>
    *
-   * <pre class="stHighlight">
+   * <pre class="stExamples">
    * they("should be empty")
    * ^
    * </pre>
    *
-   * <pre class="stHighlight">
-   * they should behave like nonFullStack(stackWithOneItem)
+   * <pre> class="stExamples"
+   * it should behave like nonFullStack(stackWithOneItem)
    * ^
    * </pre>
    *
    * <p>
-   * For more information and examples of the use of the <code>it</code> field, see
-   * the <a href="../FunSpec.html">main documentation for <code>FunSpec</code></a>.
+   * For more information and examples of the use of the <code>it</code> field, see the main documentation for this trait.
    * </p>
    */
   protected val they = new TheyWord
@@ -362,7 +353,7 @@ trait FunSpecRegistration extends Suite with TestRegistration with Informing wit
    * for <code>testNames</code> for an example.) The resulting test name must not have been registered previously on
    * this <code>FunSpec</code> instance.
    *
-   * @param specText the specification text, which will be combined with the descText of any surrounding describers
+   * @param testText the specification text, which will be combined with the descText of any surrounding describers
    * to form the test name
    * @param testTags the optional list of tags for this test
    * @param testFun the test function
@@ -370,35 +361,14 @@ trait FunSpecRegistration extends Suite with TestRegistration with Informing wit
    * @throws TestRegistrationClosedException if invoked after <code>run</code> has been invoked on this suite
    * @throws NullArgumentException if <code>specText</code> or any passed test tag is <code>null</code>
    */
-  protected def ignore(specText: String, testTags: Tag*)(testFun: FixtureParam => Registration) {
+  protected def ignore(testText: String, testTags: Tag*)(testFun: => Registration) {
     // SKIP-SCALATESTJS-START
-    val stackDepth = 6
+    val stackDepth = 4
+    val stackDepthAdjustment = -2
     // SKIP-SCALATESTJS-END
-    //SCALATESTJS-ONLY val stackDepth = 8
-    engine.registerIgnoredTest(specText, transformToOutcome(testFun), Resources.ignoreCannotAppearInsideAnItOrAThey, sourceFileName, "ignore", stackDepth, -2, None, testTags: _*)
-  }
-
-  /**
-   * Register a test to ignore, which has the given spec text and test function value that takes no arguments.
-   * This method will register the test for later ignoring via an invocation of one of the <code>execute</code>
-   * methods. This method exists to make it easy to ignore an existing test by changing the call to <code>it</code>
-   * to <code>ignore</code> without deleting or commenting out the actual test code. The test will not be executed, but a
-   * report will be sent that indicates the test was ignored. The name of the test will be a concatenation of the text of all surrounding describers,
-   * from outside in, and the passed spec text, with one space placed between each item. (See the documenation
-   * for <code>testNames</code> for an example.) The resulting test name must not have been registered previously on
-   * this <code>FunSpec</code> instance.
-   *
-   * @param specText the specification text, which will be combined with the descText of any surrounding describers
-   * to form the test name
-   * @param testFun the test function
-   * @throws DuplicateTestNameException if a test with the same name has been registered previously
-   * @throws TestRegistrationClosedException if invoked after <code>run</code> has been invoked on this suite
-   * @throws NullArgumentException if <code>specText</code> or any passed test tag is <code>null</code>
-   */
-  protected def ignore(specText: String)(testFun: FixtureParam => Registration) {
-    if (atomic.get.registrationClosed)
-      throw new TestRegistrationClosedException(Resources.ignoreCannotAppearInsideAnItOrAThey, getStackDepthFun(sourceFileName, "ignore"))
-    ignore(specText, Array[Tag](): _*)(testFun)
+    //SCALATESTJS-ONLY val stackDepth = 6
+    //SCALATESTJS-ONLY val stackDepthAdjustment = -6
+    engine.registerIgnoredTest(testText, transformToOutcome(testFun), Resources.ignoreCannotAppearInsideAnItOrAThey, sourceFileName, "ignore", stackDepth, stackDepthAdjustment, None, testTags: _*)
   }
 
   /**
@@ -406,9 +376,6 @@ trait FunSpecRegistration extends Suite with TestRegistration with Informing wit
    * passed function value may contain more describers (defined with <code>describe</code>) and/or tests
    * (defined with <code>it</code>). This trait's implementation of this method will register the
    * description string and immediately invoke the passed function.
-   *
-   * @param description the description text
-   * @param fun the function which makes up the body for the description
    */
   protected def describe(description: String)(fun: => Unit) {
     // SKIP-SCALATESTJS-START
@@ -429,20 +396,45 @@ trait FunSpecRegistration extends Suite with TestRegistration with Informing wit
   }
 
   /**
-   * A <code>Map</code> whose keys are <code>String</code> tag names to which tests in this <code>FunSpec</code> belong, and values
-   * the <code>Set</code> of test names that belong to each tag. If this <code>FunSpec</code> contains no tags, this method returns an empty <code>Map</code>.
+   * An immutable <code>Set</code> of test names. If this <code>FunSpec</code> contains no tests, this method returns an
+   * empty <code>Set</code>.
    *
    * <p>
-   * This trait's implementation returns tags that were passed as strings contained in <code>Tag</code> objects passed to
-   * methods <code>test</code> and <code>ignore</code>.
+   * This trait's implementation of this method will return a set that contains the names of all registered tests. The set's
+   * iterator will return those names in the order in which the tests were registered. Each test's name is composed
+   * of the concatenation of the text of each surrounding describer, in order from outside in, and the text of the
+   * example itself, with all components separated by a space. For example, consider this <code>FunSpec</code>:
    * </p>
    *
+   * <pre class="stHighlight">
+   * import org.scalatest.FunSpec
+   *
+   * class StackSpec extends FunSpec {
+   *   describe("A Stack") {
+   *     describe("(when not empty)") {
+   *       it("must allow me to pop") {}
+   *     }
+   *     describe("(when not full)") {
+   *       it("must allow me to push") {}
+   *     }
+   *   }
+   * }
+   * </pre>
+   *
    * <p>
-   * In addition, this trait's implementation will also auto-tag tests with class level annotations.
-   * For example, if you annotate @Ignore at the class level, all test methods in the class will be auto-annotated with @Ignore.
+   * Invoking <code>testNames</code> on this <code>FunSpec</code> will yield a set that contains the following
+   * two test name strings:
    * </p>
+   *
+   * <pre class="stExamples">
+   * "A Stack (when not empty) must allow me to pop"
+   * "A Stack (when not full) must allow me to push"
+   * </pre>
    */
-  override def tags: Map[String, Set[String]] = autoTagClassAnnotations(atomic.get.tagsMap, this)
+  override def testNames: Set[String] = {
+    // I'm returning a ListSet here so that they tests will be run in registration order
+    ListSet(atomic.get.testNamesList.toArray: _*)
+  }
 
   /**
    * Run a test. This trait's implementation runs the test registered with the name specified by
@@ -453,26 +445,26 @@ trait FunSpecRegistration extends Suite with TestRegistration with Informing wit
    * @param testName the name of one test to execute.
    * @param args the <code>Args</code> for this run
    * @return a <code>Status</code> object that indicates when the test started by this method has completed, and whether or not it failed .
-   * @throws NullArgumentException if <code>testName</code> or <code>args</code> is <code>null</code>.
+   *
+   * @throws NullArgumentException if any of <code>testName</code>, <code>reporter</code>, <code>stopper</code>, or <code>configMap</code>
+   *     is <code>null</code>.
    */
   protected override def runTest(testName: String, args: Args): Status = {
 
     def invokeWithFixture(theTest: TestLeaf): AsyncOutcome = {
+      val theConfigMap = args.configMap
+      val testData = testDataFor(testName, theConfigMap)
       PastOutcome(
-        theTest.testFun match {
-          case transformer: org.scalatest.fixture.Transformer[_] =>
-            transformer.exceptionalTestFun match {
-              case wrapper: NoArgTestWrapper[_, _] =>
-                withFixture(new FixturelessTestFunAndConfigMap(testName, wrapper.test, args.configMap))
-              case fun => withFixture(new TestFunAndConfigMap(testName, fun, args.configMap))
-            }
-          case other =>
-            other match {
-              case wrapper: NoArgTestWrapper[_, _] =>
-                withFixture(new FixturelessTestFunAndConfigMap(testName, wrapper.test, args.configMap))
-              case fun => withFixture(new TestFunAndConfigMap(testName, fun, args.configMap))
-            }
-        }
+        withFixture(
+          new NoArgTest {
+            val name = testData.name
+            def apply(): Outcome = { theTest.testFun().toOutcome }
+            val configMap = testData.configMap
+            val scopes = testData.scopes
+            val text = testData.text
+            val tags = testData.tags
+          }
+        )
       )
     }
 
@@ -480,65 +472,36 @@ trait FunSpecRegistration extends Suite with TestRegistration with Informing wit
   }
 
   /**
+   * A <code>Map</code> whose keys are <code>String</code> names of tagged tests and whose associated values are
+   * the <code>Set</code> of tags for the test. If this <code>FunSpec</code> contains no tags, this method returns an empty <code>Map</code>.
+   *
    * <p>
-   * Run zero to many of this <code>FunSpec</code>'s tests.
+   * This trait's implementation returns tags that were passed as strings contained in <code>Tag</code> objects passed to
+   * methods <code>it</code> and <code>ignore</code>.
    * </p>
    *
    * <p>
-   * This method takes a <code>testName</code> parameter that optionally specifies a test to invoke.
-   * If <code>testName</code> is <code>Some</code>, this trait's implementation of this method
-   * invokes <code>runTest</code> on this object with passed <code>args</code>.
+   * In addition, this trait's implementation will also auto-tag tests with class level annotations.
+   * For example, if you annotate <code>@Ignore</code> at the class level, all test methods in the class will be auto-annotated with
+   * <code>org.scalatest.Ignore</code>.
    * </p>
-   *
-   * <p>
-   * This method takes an <code>args</code> that contains a <code>Set</code> of tag names that should be included (<code>tagsToInclude</code>), and a <code>Set</code>
-   * that should be excluded (<code>tagsToExclude</code>), when deciding which of this <code>Suite</code>'s tests to execute.
-   * If <code>tagsToInclude</code> is empty, all tests will be executed
-   * except those those belonging to tags listed in the <code>tagsToExclude</code> <code>Set</code>. If <code>tagsToInclude</code> is non-empty, only tests
-   * belonging to tags mentioned in <code>tagsToInclude</code>, and not mentioned in <code>tagsToExclude</code>
-   * will be executed. However, if <code>testName</code> is <code>Some</code>, <code>tagsToInclude</code> and <code>tagsToExclude</code> are essentially ignored.
-   * Only if <code>testName</code> is <code>None</code> will <code>tagsToInclude</code> and <code>tagsToExclude</code> be consulted to
-   * determine which of the tests named in the <code>testNames</code> <code>Set</code> should be run. For more information on trait tags, see the main documentation for this trait.
-   * </p>
-   *
-   * <p>
-   * If <code>testName</code> is <code>None</code>, this trait's implementation of this method
-   * invokes <code>testNames</code> on this <code>Suite</code> to get a <code>Set</code> of names of tests to potentially execute.
-   * (A <code>testNames</code> value of <code>None</code> essentially acts as a wildcard that means all tests in
-   * this <code>Suite</code> that are selected by <code>tagsToInclude</code> and <code>tagsToExclude</code> should be executed.)
-   * For each test in the <code>testName</code> <code>Set</code>, in the order
-   * they appear in the iterator obtained by invoking the <code>elements</code> method on the <code>Set</code>, this trait's implementation
-   * of this method checks whether the test should be run based on the <code>tagsToInclude</code> and <code>tagsToExclude</code> <code>Set</code>s.
-   * If so, this implementation invokes <code>runTest</code> with passed <code>args</code>.
-   * </p>
-   *
-   * @param testName an optional name of one test to execute. If <code>None</code>, all relevant tests should be executed.
-   *                 I.e., <code>None</code> acts like a wildcard that means execute all relevant tests in this <code>FunSpec</code>.
-   * @param args the <code>Args</code> to which results will be reported
-   * @return a <code>Status</code> object that indicates when all tests started by this method have completed, and whether or not a failure occurred.
-   * @throws NullArgumentException if any of <code>testName</code> or <code>args</code> is <code>null</code>.
    */
-  protected override def runTests(testName: Option[String], args: Args): Status = {
-
-    runTestsImpl(thisSuite, testName, args, info, true, runTest)
-  }
+  override def tags: Map[String, Set[String]] = autoTagClassAnnotations(atomic.get.tagsMap, this)
 
   /**
-   * An immutable <code>Set</code> of test names. If this <code>FunSpec</code> contains no tests, this method returns an
-   * empty <code>Set</code>.
+   * Run zero to many of this <code>FunSpec</code>'s tests.
    *
-   * <p>
-   * This trait's implementation of this method will return a set that contains the names of all registered tests. The set's
-   * iterator will return those names in the order in which the tests were registered. Each test's name is composed
-   * of the concatenation of the text of each surrounding describer, in order from outside in, and the text of the
-   * example itself, with all components separated by a space.
-   * </p>
+   * @param testName an optional name of one test to run. If <code>None</code>, all relevant tests should be run.
+   *                 I.e., <code>None</code> acts like a wildcard that means run all relevant tests in this <code>Suite</code>.
+   * @param args the <code>Args</code> for this run
+   * @return a <code>Status</code> object that indicates when all tests started by this method have completed, and whether or not a failure occurred.
    *
-   * @return the <code>Set</code> of test names
+   * @throws NullArgumentException if any of the passed parameters is <code>null</code>.
+   * @throws IllegalArgumentException if <code>testName</code> is defined, but no test with the specified test name
+   *     exists in this <code>Suite</code>
    */
-  override def testNames: Set[String] = {
-    // I'm returning a ListSet here so that they tests will be run in registration order
-    ListSet(atomic.get.testNamesList.toArray: _*)
+  protected override def runTests(testName: Option[String], args: Args): Status = {
+    runTestsImpl(thisSuite, testName, args, info, true, runTest)
   }
 
   override def run(testName: Option[String], args: Args): Status = {
@@ -552,54 +515,22 @@ trait FunSpecRegistration extends Suite with TestRegistration with Informing wit
    * This field supports syntax such as the following:
    * </p>
    *
-   * <pre class="stHighlight">
+   * <pre class="stExamples">
    * it should behave like nonFullStack(stackWithOneItem)
    *           ^
    * </pre>
    *
    * <p>
-   * For more information and examples of the use of <cod>behave</code>, see the <a href="../FunSpec.html#SharedTests">Shared tests section</a>
-   * in the main documentation for trait <code>FunSpec</code>.
+   * For more information and examples of the use of <cod>behave</code>, see the <a href="#sharedTests">Shared tests section</a>
+   * in the main documentation for this trait.
    * </p>
    */
   protected val behave = new BehaveWord
 
-  import scala.language.implicitConversions
-
-  /**
-   * Implicitly converts a function that takes no parameters and results in <code>PendingNothing</code> to
-   * a function from <code>FixtureParam</code> to <code>Any</code>, to enable pending tests to registered as by-name parameters
-   * by methods that require a test function that takes a <code>FixtureParam</code>.
-   *
-   * <p>
-   * This method makes it possible to write pending tests as simply <code>(pending)</code>, without needing
-   * to write <code>(fixture => pending)</code>.
-   * </p>
-   *
-   * @param f a function
-   * @return a function of <code>FixtureParam => Any</code>
-   */
-  protected implicit def convertPendingToFixtureFunction(f: => PendingNothing): FixtureParam => Any = {
-    fixture => f
-  }
-
-  /**
-   * Implicitly converts a function that takes no parameters and results in <code>Any</code> to
-   * a function from <code>FixtureParam</code> to <code>Any</code>, to enable no-arg tests to registered
-   * by methods that require a test function that takes a <code>FixtureParam</code>.
-   *
-   * @param fun a function
-   * @return a function of <code>FixtureParam => Any</code>
-   */
-  protected implicit def convertNoArgToFixtureFunction(fun: () => Any): (FixtureParam => Any) =
-    new NoArgTestWrapper(fun)
-
   /**
    * Suite style name.
-   *
-   * @return <code>org.scalatest.fixture.FunSpec</code>
    */
-  final override val styleName: String = "org.scalatest.fixture.FunSpec"
+  final override val styleName: String = "org.scalatest.FunSpec"
 
   override def testDataFor(testName: String, theConfigMap: ConfigMap = ConfigMap.empty): TestData = createTestDataFor(testName, theConfigMap, this)
 }
