+++ scalatest/src/main/scala/org/scalatest/PropSpecLike.scala
@@ -16,4 +16,4 @@
 package org.scalatest
 
 //SCALATESTJS-ONLY @scala.scalajs.js.annotation.JSExportDescendentClasses(ignoreInvalidDescendants = true)
-trait PropSpecLike extends PropSpecRegistration with ClassicTests
+trait PropSpecLike extends PropSpecOf[Unit]
