+++ scalatest/src/main/scala/org/scalatest/fixture/AsyncFeatureSpecLike.scala
@@ -19,7 +19,7 @@ import scala.concurrent.{Future, ExecutionContext}
 import org.scalatest._
 
 //SCALATESTJS-ONLY @scala.scalajs.js.annotation.JSExportDescendentClasses(ignoreInvalidDescendants = true)
-trait AsyncFeatureSpecLike extends FeatureSpecRegistration with AsyncTests with org.scalatest.OneInstancePerTest { thisSuite =>
+trait AsyncFeatureSpecLike extends FeatureSpecOf[Future[Any]] with AsyncFixtures with org.scalatest.OneInstancePerTest { thisSuite =>
 
   implicit def executionContext: ExecutionContext
