+++ scalatest/src/main/scala/org/scalatest/fixture/AsyncPropSpecLike.scala
@@ -19,7 +19,7 @@ import scala.concurrent.{Future, ExecutionContext}
 import org.scalatest._
 
 //SCALATESTJS-ONLY @scala.scalajs.js.annotation.JSExportDescendentClasses(ignoreInvalidDescendants = true)
-trait AsyncPropSpecLike extends PropSpecRegistration with AsyncTests with org.scalatest.OneInstancePerTest { thisSuite =>
+trait AsyncPropSpecLike extends PropSpecOf[Future[Any]] with AsyncFixtures with org.scalatest.OneInstancePerTest { thisSuite =>
 
   implicit def executionContext: ExecutionContext
