+++ scalatest/src/main/scala/org/scalatest/fixture/FeatureSpecLike.scala
@@ -16,4 +16,4 @@
 package org.scalatest.fixture
 
 //SCALATESTJS-ONLY @scala.scalajs.js.annotation.JSExportDescendentClasses(ignoreInvalidDescendants = true)
-trait FeatureSpecLike extends FeatureSpecRegistration with ClassicTests
+trait FeatureSpecLike extends FeatureSpecOf[Any]
