+++ scalatest/src/main/scala/org/scalatest/fixture/FeatureSpecOf.scala
@@ -13,46 +13,51 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest
+package org.scalatest.fixture
 
+import org.scalatest._
 import scala.collection.immutable.ListSet
 import org.scalatest.exceptions.StackDepthExceptionHelper.getStackDepthFun
 import java.util.concurrent.atomic.AtomicReference
 import java.util.ConcurrentModificationException
 import org.scalatest.events._
-import Suite.anExceptionThatShouldCauseAnAbort
-import Suite.autoTagClassAnnotations
+import org.scalatest.Suite.anExceptionThatShouldCauseAnAbort
+import org.scalatest.Suite.autoTagClassAnnotations
 import org.scalatest.exceptions.NotAllowedException
 
 /**
- * Implementation trait for class <code>FeatureSpec</code>, which represents
- * a suite of tests in which each test represents one <em>scenario</em> of a
- * <em>feature</em>.
+ * Implementation trait for class <code>fixture.FeatureSpec</code>, which is
+ * a sister class to <a href="../FeatureSpec.html"><code>org.scalatest.FeatureSpec</code></a> that can pass a
+ * fixture object into its tests.
  *
  * <p>
- * <a href="FeatureSpec.html"><code>FeatureSpec</code></a> is a class, not a
- * trait, to minimize compile time given there is a slight compiler overhead to
- * mixing in traits compared to extending classes. If you need to mix the
- * behavior of <code>FeatureSpec</code> into some other class, you can use this
- * trait instead, because class <code>FeatureSpec</code> does nothing more than
- * extend this trait and add a nice <code>toString</code> implementation.
+ * <a href="FeatureSpec.html"><code>fixture.FeatureSpec</code></a> is a class,
+ * not a trait, to minimize compile time given there is a slight compiler
+ * overhead to mixing in traits compared to extending classes. If you need
+ * to mix the behavior of <code>fixture.FeatureSpec</code> into some other
+ * class, you can use this trait instead, because class
+ * <code>fixture.FeatureSpec</code> does nothing more than extend this trait and add a nice <code>toString</code> implementation.
  * </p>
  *
  * <p>
  * See the documentation of the class for a <a href="FeatureSpec.html">detailed
- * overview of <code>FeatureSpec</code></a>.
+ * overview of <code>fixture.FeatureSpec</code></a>.
  * </p>
  *
  * @author Bill Venners
  */
 @Finders(Array("org.scalatest.finders.FeatureSpecFinder"))
-trait FeatureSpecRegistration extends Suite with TestRegistration with Informing with Notifying with Alerting with Documenting { thisSuite =>
+trait FeatureSpecOf[R] extends Suite with TestRegistration with Informing with Notifying with Alerting with Documenting { thisSuite =>
 
-  private final val engine = new Engine(Resources.concurrentFeatureSpecMod, "FeatureSpec")
+  type Registration = R
 
-  protected[scalatest] def getEngine: Engine = engine
+  private final val engine = new FixtureEngine[FixtureParam](Resources.concurrentFeatureSpecMod, "FixtureFeatureSpec")
+
+  protected[scalatest] def getEngine: FixtureEngine[FixtureParam] = engine
 
   import engine._
+
+  private[scalatest] val sourceFileName = "FeatureSpecOf.scala"
 
   /**
    * Returns an <code>Informer</code> that during test execution will forward strings passed to its
@@ -96,20 +101,20 @@ trait FeatureSpecRegistration extends Suite with TestRegistration with Informing
    */
   protected def markup: Documenter = atomicDocumenter.get
 
-  final def registerTest(testText: String, testTags: Tag*)(testFun: => Registration) {
+  final def registerTest(testText: String, testTags: Tag*)(testFun: FixtureParam => Registration) {
     // SKIP-SCALATESTJS-START
     val stackDepthAdjustment = -1
     // SKIP-SCALATESTJS-END
     //SCALATESTJS-ONLY val stackDepthAdjustment = -4
-    engine.registerTest(Resources.scenario(testText.trim), transformToOutcome(testFun), Resources.testCannotBeNestedInsideAnotherTest, "FeatureSpecRegistration.scala", "registerTest", 4, stackDepthAdjustment, None, None, None, testTags: _*)
+    engine.registerTest(Resources.scenario(testText.trim), transformToOutcome(testFun), Resources.testCannotBeNestedInsideAnotherTest, "FeatureSpecOf.scala", "registerTest", 4, stackDepthAdjustment, None, None, None, testTags: _*)
   }
 
-  final def registerIgnoredTest(testText: String, testTags: Tag*)(testFun: => Registration) {
+  final def registerIgnoredTest(testText: String, testTags: Tag*)(testFun: FixtureParam => Registration) {
     // SKIP-SCALATESTJS-START
-    val stackDepthAdjustment = -2
+    val stackDepthAdjustment = -3
     // SKIP-SCALATESTJS-END
     //SCALATESTJS-ONLY val stackDepthAdjustment = -5
-    engine.registerIgnoredTest(Resources.scenario(testText.trim), transformToOutcome(testFun), Resources.testCannotBeNestedInsideAnotherTest, "FeatureSpecRegistration.scala", "registerIgnoredTest", 4, stackDepthAdjustment, None, testTags: _*)
+    engine.registerIgnoredTest(Resources.scenario(testText.trim), transformToOutcome(testFun), Resources.testCannotBeNestedInsideAnotherTest, "FeatureSpecOf.scala", "registerIgnoredTest", 4, stackDepthAdjustment, None, testTags: _*)
   }
 
   /**
@@ -130,14 +135,14 @@ trait FeatureSpecRegistration extends Suite with TestRegistration with Informing
    * @throws TestRegistrationClosedException if invoked after <code>run</code> has been invoked on this suite
    * @throws NullArgumentException if <code>specText</code> or any passed test tag is <code>null</code>
    */
-  protected def scenario(specText: String, testTags: Tag*)(testFun: => Registration) {
+  protected def scenario(specText: String, testTags: Tag*)(testFun: FixtureParam => Registration) {
     // SKIP-SCALATESTJS-START
     val stackDepth = 4
     val stackDepthAdjustment = -2
     // SKIP-SCALATESTJS-END
     //SCALATESTJS-ONLY val stackDepth = 6
-    //SCALATESTJS-ONLY val stackDepthAdjustment = -4
-    engine.registerTest(Resources.scenario(specText.trim), transformToOutcome(testFun), Resources.scenarioCannotAppearInsideAnotherScenario, "FeatureSpecRegistration.scala", "scenario", stackDepth, stackDepthAdjustment, None, None, None, testTags: _*)
+    //SCALATESTJS-ONLY val stackDepthAdjustment = -6
+    engine.registerTest(Resources.scenario(specText.trim), transformToOutcome(testFun), Resources.scenarioCannotAppearInsideAnotherScenario, sourceFileName, "scenario", stackDepth, stackDepthAdjustment, None, None, None, testTags: _*)
   }
 
   /**
@@ -158,14 +163,14 @@ trait FeatureSpecRegistration extends Suite with TestRegistration with Informing
    * @throws TestRegistrationClosedException if invoked after <code>run</code> has been invoked on this suite
    * @throws NullArgumentException if <code>specText</code> or any passed test tag is <code>null</code>
    */
-  protected def ignore(specText: String, testTags: Tag*)(testFun: => Registration) {
+  protected def ignore(specText: String, testTags: Tag*)(testFun: FixtureParam => Registration) {
     // SKIP-SCALATESTJS-START
     val stackDepth = 4
-    val stackDepthAdjustment = -2
+    val stackDepthAdjustment = -3
     // SKIP-SCALATESTJS-END
     //SCALATESTJS-ONLY val stackDepth = 6
-    //SCALATESTJS-ONLY val stackDepthAdjustment = -5
-    engine.registerIgnoredTest(Resources.scenario(specText), transformToOutcome(testFun), Resources.ignoreCannotAppearInsideAScenario, "FeatureSpecRegistration.scala", "ignore", stackDepth, stackDepthAdjustment, None, testTags: _*)
+    //SCALATESTJS-ONLY val stackDepthAdjustment = -7
+    engine.registerIgnoredTest(Resources.scenario(specText), transformToOutcome(testFun), Resources.ignoreCannotAppearInsideAScenario, sourceFileName, "ignore", stackDepth, stackDepthAdjustment, None, testTags: _*)
   }
 
   /**
@@ -173,6 +178,8 @@ trait FeatureSpecRegistration extends Suite with TestRegistration with Informing
    * passed function value may contain more describers (defined with <code>describe</code>) and/or tests
    * (defined with <code>it</code>). This trait's implementation of this method will register the
    * description string and immediately invoke the passed function.
+   *
+   * @param description the description text
    */
   protected def feature(description: String)(fun: => Unit) {
 
@@ -186,10 +193,10 @@ trait FeatureSpecRegistration extends Suite with TestRegistration with Informing
     //SCALATESTJS-ONLY val scopeErrorStackDepth = 11
 
     if (!currentBranchIsTrunk)
-      throw new NotAllowedException(Resources.cantNestFeatureClauses, getStackDepthFun("FeatureSpecRegistration.scala", "feature"))
+      throw new NotAllowedException(Resources.cantNestFeatureClauses, getStackDepthFun(sourceFileName, "feature"))
 
     try {
-      registerNestedBranch(Resources.feature(description.trim), None, fun, Resources.featureCannotAppearInsideAScenario, "FeatureSpecRegistration.scala", "feature", stackDepth, stackDepthAdjustment, None)
+      registerNestedBranch(Resources.feature(description.trim), None, fun, Resources.featureCannotAppearInsideAScenario, sourceFileName, "feature", stackDepth, stackDepthAdjustment, None)
     }
     catch {
       case e: exceptions.TestFailedException => throw new exceptions.NotAllowedException(FailureMessages.assertionShouldBePutInsideScenarioClauseNotFeatureClause, Some(e), e => scopeErrorStackDepth)
@@ -201,18 +208,17 @@ trait FeatureSpecRegistration extends Suite with TestRegistration with Informing
   }
 
   /**
-   * A <code>Map</code> whose keys are <code>String</code> names of tagged tests and whose associated values are
-   * the <code>Set</code> of tag names for the test. If this <code>FeatureSpec</code> contains no tags, this method returns an empty <code>Map</code>.
+   * A <code>Map</code> whose keys are <code>String</code> tag names to which tests in this <code>FeatureSpec</code> belong, and values
+   * the <code>Set</code> of test names that belong to each tag. If this <code>FeatureSpec</code> contains no tags, this method returns an empty <code>Map</code>.
    *
    * <p>
    * This trait's implementation returns tags that were passed as strings contained in <code>Tag</code> objects passed to
-   * methods <code>scenario</code> and <code>ignore</code>.
+   * methods <code>test</code> and <code>ignore</code>.
    * </p>
    *
    * <p>
    * In addition, this trait's implementation will also auto-tag tests with class level annotations.
-   * For example, if you annotate <code>@Ignore</code> at the class level, all test methods in the class will be auto-annotated with
-   * <code>org.scalatest.Ignore</code>.
+   * For example, if you annotate @Ignore at the class level, all test methods in the class will be auto-annotated with @Ignore.
    * </p>
    */
   override def tags: Map[String, Set[String]] = autoTagClassAnnotations(atomic.get.tagsMap, this)
@@ -226,26 +232,29 @@ trait FeatureSpecRegistration extends Suite with TestRegistration with Informing
    * @param testName the name of one test to execute.
    * @param args the <code>Args</code> for this run
    * @return a <code>Status</code> object that indicates when the test started by this method has completed, and whether or not it failed .
-   *
-   * @throws NullArgumentException if any of <code>testName</code>, <code>reporter</code>, <code>stopper</code>, or <code>configMap</code>
+   * @throws NullArgumentException if <code>testName</code>, <code>reporter</code>, <code>stopper</code>, or <code>configMap</code>
    *     is <code>null</code>.
    */
   protected override def runTest(testName: String, args: Args): Status = {
 
+
     def invokeWithFixture(theTest: TestLeaf): AsyncOutcome = {
-      val theConfigMap = args.configMap
-      val testData = testDataFor(testName, theConfigMap)
+
       PastOutcome(
-        withFixture(
-          new NoArgTest {
-            val name = testData.name
-            def apply(): Outcome = { theTest.testFun().toOutcome }
-            val configMap = testData.configMap
-            val scopes = testData.scopes
-            val text = testData.text
-            val tags = testData.tags
-          }
-        )
+        theTest.testFun match {
+          case transformer: org.scalatest.fixture.Transformer[_] =>
+            transformer.exceptionalTestFun match {
+              case wrapper: NoArgTestWrapper[_, _] =>
+                withFixture(new FixturelessTestFunAndConfigMap(testName, wrapper.test, args.configMap))
+              case fun => withFixture(new TestFunAndConfigMap(testName, fun, args.configMap))
+            }
+          case other =>
+            other match {
+              case wrapper: NoArgTestWrapper[_, _] =>
+                withFixture(new FixturelessTestFunAndConfigMap(testName, wrapper.test, args.configMap))
+              case fun => withFixture(new TestFunAndConfigMap(testName, fun, args.configMap))
+            }
+        }
       )
     }
 
@@ -253,24 +262,18 @@ trait FeatureSpecRegistration extends Suite with TestRegistration with Informing
   }
 
   /**
+   * <p>
    * Run zero to many of this <code>FeatureSpec</code>'s tests.
+   * </p>
    *
    * <p>
    * This method takes a <code>testName</code> parameter that optionally specifies a test to invoke.
    * If <code>testName</code> is <code>Some</code>, this trait's implementation of this method
-   * invokes <code>runTest</code> on this object, passing in:
+   * invokes <code>runTest</code> on this object with passed <code>args</code>.
    * </p>
    *
-   * <ul>
-   * <li><code>testName</code> - the <code>String</code> value of the <code>testName</code> <code>Option</code> passed
-   *   to this method</li>
-   * <li><code>reporter</code> - the <code>Reporter</code> passed to this method, or one that wraps and delegates to it</li>
-   * <li><code>stopper</code> - the <code>Stopper</code> passed to this method, or one that wraps and delegates to it</li>
-   * <li><code>configMap</code> - the <code>configMap</code> passed to this method, or one that wraps and delegates to it</li>
-   * </ul>
-   *
    * <p>
-   * This method takes a <code>Set</code> of tag names that should be included (<code>tagsToInclude</code>), and a <code>Set</code>
+   * This method takes an <code>args</code> that contains a <code>Set</code> of tag names that should be included (<code>tagsToInclude</code>), and a <code>Set</code>
    * that should be excluded (<code>tagsToExclude</code>), when deciding which of this <code>Suite</code>'s tests to execute.
    * If <code>tagsToInclude</code> is empty, all tests will be executed
    * except those those belonging to tags listed in the <code>tagsToExclude</code> <code>Set</code>. If <code>tagsToInclude</code> is non-empty, only tests
@@ -288,24 +291,14 @@ trait FeatureSpecRegistration extends Suite with TestRegistration with Informing
    * For each test in the <code>testName</code> <code>Set</code>, in the order
    * they appear in the iterator obtained by invoking the <code>elements</code> method on the <code>Set</code>, this trait's implementation
    * of this method checks whether the test should be run based on the <code>tagsToInclude</code> and <code>tagsToExclude</code> <code>Set</code>s.
-   * If so, this implementation invokes <code>runTest</code>, passing in:
+   * If so, this implementation invokes <code>runTest</code> with passed in <code>args</code>.
    * </p>
    *
-   * <ul>
-   * <li><code>testName</code> - the <code>String</code> name of the test to run (which will be one of the names in the <code>testNames</code> <code>Set</code>)</li>
-   * <li><code>reporter</code> - the <code>Reporter</code> passed to this method, or one that wraps and delegates to it</li>
-   * <li><code>stopper</code> - the <code>Stopper</code> passed to this method, or one that wraps and delegates to it</li>
-   * <li><code>configMap</code> - the <code>configMap</code> passed to this method, or one that wraps and delegates to it</li>
-   * </ul>
-   *
-   * @param testName an optional name of one test to run. If <code>None</code>, all relevant tests should be run.
-   *                 I.e., <code>None</code> acts like a wildcard that means run all relevant tests in this <code>Suite</code>.
+   * @param testName an optional name of one test to execute. If <code>None</code>, all relevant tests should be executed.
+   *                 I.e., <code>None</code> acts like a wildcard that means execute all relevant tests in this <code>fixture.FeatureSpec</code>.
    * @param args the <code>Args</code> for this run
    * @return a <code>Status</code> object that indicates when all tests started by this method have completed, and whether or not a failure occurred.
-   *
-   * @throws NullArgumentException if any of the passed parameters is <code>null</code>.
-   * @throws IllegalArgumentException if <code>testName</code> is defined, but no test with the specified test name
-   *     exists in this <code>Suite</code>
+   * @throws NullArgumentException if any of <code>testName</code> or <code>args</code> is <code>null</code>.
    */
   protected override def runTests(testName: Option[String], args: Args): Status = {
     runTestsImpl(thisSuite, testName, args, info, false, runTest)
@@ -319,31 +312,12 @@ trait FeatureSpecRegistration extends Suite with TestRegistration with Informing
    * This trait's implementation of this method will return a set that contains the names of all registered tests. The set's
    * iterator will return those names in the order in which the tests were registered. Each test's name is composed
    * of the concatenation of the text of each surrounding describer, in order from outside in, and the text of the
-   * example itself, with all components separated by a space. For example, consider this <code>FeatureSpec</code>:
+   * example itself, with all components separated by a space.
    * </p>
    *
-   * <pre class="stHighlight">
-   * import org.scalatest.FeatureSpec
-   *
-   * class StackSpec extends FeatureSpec {
-   *   feature("A Stack") {
-   *     scenario("(when not empty) must allow me to pop") {}
-   *     scenario("(when not full) must allow me to push") {}
-   *   }
-   * }
-   * </pre>
-   *
-   * <p>
-   * Invoking <code>testNames</code> on this <code>FeatureSpec</code> will yield a set that contains the following
-   * two test name strings:
-   * </p>
-   *
-   * <pre>
-   * "A Stack (when not empty) must allow me to pop"
-   * "A Stack (when not full) must allow me to push"
-   * </pre>
+   * @return the <code>Set</code> of test names
    */
-  // override def testNames: Set[String] = ListSet(atomic.get.testsList.map(_.testName): _*)
+  //override def testNames: Set[String] = ListSet(atomic.get.testsList.map(_.testName): _*)
   override def testNames: Set[String] = {
     // I'm returning a ListSet here so that they tests will be run in registration order
     ListSet(atomic.get.testNamesList.toArray: _*)
@@ -369,15 +343,51 @@ trait FeatureSpecRegistration extends Suite with TestRegistration with Informing
    * Because the parameter passed to it is
    * type <code>Unit</code>, the expression will be evaluated before being passed, which
    * is sufficient to register the shared scenarios. For examples of shared scenarios, see the
-   * <a href="#sharedScenarios">Shared scenarios section</a> in the main documentation for this trait.
+   * <a href="../FeatureSpec.html#SharedScenarios">Shared scenarios section</a> in the main documentation for
+   * trait <code>FeatureSpec</code>.
    * </p>
    */
   protected def scenariosFor(unit: Unit) {}
 
+  import scala.language.implicitConversions
+
+  /**
+   * Implicitly converts a function that takes no parameters and results in <code>PendingNothing</code> to
+   * a function from <code>FixtureParam</code> to <code>Any</code>, to enable pending tests to registered as by-name parameters
+   * by methods that require a test function that takes a <code>FixtureParam</code>.
+   *
+   * <p>
+   * This method makes it possible to write pending tests as simply <code>(pending)</code>, without needing
+   * to write <code>(fixture => pending)</code>.
+   * </p>
+   *
+   * @param f a function
+   * @return a function of <code>FixtureParam => Any</code>
+   */
+  protected implicit def convertPendingToFixtureFunction(f: => PendingNothing): FixtureParam => Any = {
+    fixture => f
+  }
+
+  // I need this implicit because the function is passed to scenario as the 2nd parameter list, and
+  // I can't overload on that. I could if I took the ScenarioWord approach, but that has possibly a worse
+  // downside of people could just say scenario("...") and nothing else.
+  /**
+   * Implicitly converts a function that takes no parameters and results in <code>Any</code> to
+   * a function from <code>FixtureParam</code> to <code>Any</code>, to enable no-arg tests to registered
+   * by methods that require a test function that takes a <code>FixtureParam</code>.
+   *
+   * @param fun a function
+   * @return a function of <code>FixtureParam => Any</code>
+   */
+  protected implicit def convertNoArgToFixtureFunction(fun: () => Any): (FixtureParam => Any) =
+    new NoArgTestWrapper(fun)
+
   /**
    * Suite style name.
+   *
+   * @return <code>org.scalatest.fixture.FeatureSpec</code>
    */
-  final override val styleName: String = "org.scalatest.FeatureSpec"
+  final override val styleName: String = "org.scalatest.fixture.FeatureSpec"
 
   override def testDataFor(testName: String, theConfigMap: ConfigMap = ConfigMap.empty): TestData = createTestDataFor(testName, theConfigMap, this)
 }
