+++ scalatest/src/main/scala/org/scalatest/fixture/FreeSpecLike.scala
@@ -16,4 +16,4 @@
 package org.scalatest.fixture
 
 //SCALATESTJS-ONLY @scala.scalajs.js.annotation.JSExportDescendentClasses(ignoreInvalidDescendants = true)
-trait FreeSpecLike extends FreeSpecRegistration with ClassicTests
+trait FreeSpecLike extends FreeSpecOf[Any]
