+++ scalatest/src/main/scala/org/scalatest/fixture/FunSuiteLike.scala
@@ -16,4 +16,4 @@
 package org.scalatest.fixture
 
 //SCALATESTJS-ONLY @scala.scalajs.js.annotation.JSExportDescendentClasses(ignoreInvalidDescendants = true)
-trait FunSuiteLike extends FunSuiteRegistration with ClassicTests
+trait FunSuiteLike extends FunSuiteOf[Any]
