+++ scalatest/src/main/scala/org/scalatest/fixture/PropSpecOf.scala
@@ -20,36 +20,38 @@ import scala.collection.immutable.ListSet
 import org.scalatest.Suite.autoTagClassAnnotations
 
 /**
- * Implementation trait for class <code>fixture.FunSuite</code>, which is
- * a sister class to <code>org.scalatest.FunSuite</code> that can pass a
+ * Implementation trait for class <code>fixture.PropSpec</code>, which is
+ * a sister class to <a href="../PropSpec.html"><code>org.scalatest.PropSpec</code></a> that can pass a
  * fixture object into its tests.
  *
  * <p>
- * <a href="FunSuite.html"><code>fixture.FunSuite</code></a> is a class,
+ * <a href="PropSpec.html"><code>fixture.PropSpec</code></a> is a class,
  * not a trait, to minimize compile time given there is a slight compiler
  * overhead to mixing in traits compared to extending classes. If you need
- * to mix the behavior of <code>fixture.FunSuite</code> into some other
+ * to mix the behavior of <code>fixture.PropSpec</code> into some other
  * class, you can use this trait instead, because class
- * <code>fixture.FunSuite</code> does nothing more than extend this trait and add a nice <code>toString</code> implementation.
+ * <code>fixture.PropSpec</code> does nothing more than extend this trait and add a nice <code>toString</code> implementation.
  * </p>
  *
  * <p>
- * See the documentation of the class for a <a href="FunSuite.html">detailed
- * overview of <code>fixture.FunSuite</code></a>.
+ * See the documentation of the class for a <a href="PropSpec.html">detailed
+ * overview of <code>fixture.PropSpec</code></a>.
  * </p>
  *
  * @author Bill Venners
  */
-@Finders(Array("org.scalatest.finders.FunSuiteFinder"))
-trait FunSuiteRegistration extends Suite with TestRegistration with Informing with Notifying with Alerting with Documenting { thisSuite =>
+@Finders(Array("org.scalatest.finders.PropSpecFinder"))
+trait PropSpecOf[R] extends Suite with TestRegistration with Informing with Notifying with Alerting with Documenting { thisSuite =>
 
-  private final val engine = new FixtureEngine[FixtureParam](Resources.concurrentFixtureFunSuiteMod, "FixtureFunSuite")
+  type Registration = R
+
+  private final val engine = new FixtureEngine[FixtureParam](Resources.concurrentFixturePropSpecMod, "FixturePropSpec")
 
   protected[scalatest] def getEngine: FixtureEngine[FixtureParam] = engine
 
   import engine._
 
-  private[scalatest] val sourceFileName = "FunSuiteRegistration.scala"
+  private[scalatest] val sourceFileName = "PropSpecOf.scala"
 
   /**
    * Returns an <code>Informer</code> that during test execution will forward strings passed to its
@@ -66,7 +68,7 @@ trait FunSuiteRegistration extends Suite with TestRegistration with Informing wi
    * Returns a <code>Notifier</code> that during test execution will forward strings (and other objects) passed to its
    * <code>apply</code> method to the current reporter. If invoked in a constructor, it
    * will register the passed string for forwarding later during test execution. If invoked while this
-   * <code>fixture.FunSuite</code> is being executed, such as from inside a test function, it will forward the information to
+   * <code>fixture.PropSpec</code> is being executed, such as from inside a test function, it will forward the information to
    * the current reporter immediately. If invoked at any other time, it will
    * print to the standard output. This method can be called safely by any thread.
    */
@@ -76,7 +78,7 @@ trait FunSuiteRegistration extends Suite with TestRegistration with Informing wi
    * Returns an <code>Alerter</code> that during test execution will forward strings (and other objects) passed to its
    * <code>apply</code> method to the current reporter. If invoked in a constructor, it
    * will register the passed string for forwarding later during test execution. If invoked while this
-   * <code>fixture.FunSuite</code> is being executed, such as from inside a test function, it will forward the information to
+   * <code>fixture.PropSpec</code> is being executed, such as from inside a test function, it will forward the information to
    * the current reporter immediately. If invoked at any other time, it will
    * print to the standard output. This method can be called safely by any thread.
    */
@@ -98,22 +100,22 @@ trait FunSuiteRegistration extends Suite with TestRegistration with Informing wi
     val stackDepthAdjustment = -1
     // SKIP-SCALATESTJS-END
     //SCALATESTJS-ONLY val stackDepthAdjustment = -4
-    engine.registerTest(testText, transformToOutcome(testFun), Resources.testCannotBeNestedInsideAnotherTest, "FunSuite.scala", "registerTest", 4, stackDepthAdjustment, None, None, None, testTags: _*)
+    engine.registerTest(testText, transformToOutcome(testFun), Resources.testCannotBeNestedInsideAnotherTest, sourceFileName, "registerTest", 4, stackDepthAdjustment, None, None, None, testTags: _*)
   }
 
   final def registerIgnoredTest(testText: String, testTags: Tag*)(testFun: FixtureParam => Registration) {
     // SKIP-SCALATESTJS-START
-    val stackDepthAdjustment = -4
+    val stackDepthAdjustment = -3
     // SKIP-SCALATESTJS-END
     //SCALATESTJS-ONLY val stackDepthAdjustment = -5
-    engine.registerIgnoredTest(testText, transformToOutcome(testFun), Resources.testCannotBeNestedInsideAnotherTest, "FunSuite.scala", "registerIgnoredTest", 4, stackDepthAdjustment, None, testTags: _*)
+    engine.registerIgnoredTest(testText, transformToOutcome(testFun), Resources.testCannotBeNestedInsideAnotherTest, sourceFileName, "registerIgnoredTest", 4, stackDepthAdjustment, None, testTags: _*)
   }
 
   /**
-   * Register a test with the specified name, optional tags, and function value that takes no arguments.
+   * Register a property-based test with the specified name, optional tags, and function value that takes no arguments.
    * This method will register the test for later execution via an invocation of one of the <code>run</code>
    * methods. The passed test name must not have been registered previously on
-   * this <code>FunSuite</code> instance.
+   * this <code>PropSpec</code> instance.
    *
    * @param testName the name of the test
    * @param testTags the optional list of tags for this test
@@ -123,23 +125,23 @@ trait FunSuiteRegistration extends Suite with TestRegistration with Informing wi
    * @throws NotAllowedException if <code>testName</code> had been registered previously
    * @throws NullArgumentException if <code>testName</code> or any passed test tag is <code>null</code>
    */
-  protected def test(testName: String, testTags: Tag*)(testFun: FixtureParam => Registration) {
+  protected def property(testName: String, testTags: Tag*)(testFun: FixtureParam => Registration) {
     // SKIP-SCALATESTJS-START
     val stackDepth = 4
     val stackDepthAdjustment = -2
     // SKIP-SCALATESTJS-END
     //SCALATESTJS-ONLY val stackDepth = 6
     //SCALATESTJS-ONLY val stackDepthAdjustment = -6
-    engine.registerTest(testName, transformToOutcome(testFun), Resources.testCannotAppearInsideAnotherTest, sourceFileName, "test", stackDepth, stackDepthAdjustment, None, None, None, testTags: _*)
+    engine.registerTest(testName, transformToOutcome(testFun), Resources.propertyCannotAppearInsideAnotherProperty, sourceFileName, "property", stackDepth, stackDepthAdjustment, None, None, None, testTags: _*)
   }
 
   /**
-   * Register a test to ignore, which has the specified name, optional tags, and function value that takes no arguments.
+   * Register a property-based test to ignore, which has the specified name, optional tags, and function value that takes no arguments.
    * This method will register the test for later ignoring via an invocation of one of the <code>run</code>
    * methods. This method exists to make it easy to ignore an existing test by changing the call to <code>test</code>
    * to <code>ignore</code> without deleting or commenting out the actual test code. The test will not be run, but a
    * report will be sent that indicates the test was ignored. The passed test name must not have been registered previously on
-   * this <code>FunSuite</code> instance.
+   * this <code>PropSpec</code> instance.
    *
    * @param testName the name of the test
    * @param testTags the optional list of tags for this test
@@ -155,11 +157,11 @@ trait FunSuiteRegistration extends Suite with TestRegistration with Informing wi
     // SKIP-SCALATESTJS-END
     //SCALATESTJS-ONLY val stackDepth = 6
     //SCALATESTJS-ONLY val stackDepthAdjustment = -7
-    engine.registerIgnoredTest(testName, transformToOutcome(testFun), Resources.ignoreCannotAppearInsideATest, sourceFileName, "ignore", stackDepth, stackDepthAdjustment, None, testTags: _*)
+    engine.registerIgnoredTest(testName, transformToOutcome(testFun), Resources.ignoreCannotAppearInsideAProperty, sourceFileName, "ignore", stackDepth, stackDepthAdjustment, None, testTags: _*)
   }
 
   /**
-   * An immutable <code>Set</code> of test names. If this <code>fixture.FunSuite</code> contains no tests, this method returns an empty <code>Set</code>.
+   * An immutable <code>Set</code> of test names. If this <code>fixture.PropSpec</code> contains no tests, this method returns an empty <code>Set</code>.
    *
    * <p>
    * This trait's implementation of this method will return a set that contains the names of all registered tests. The set's iterator will
@@ -179,8 +181,8 @@ trait FunSuiteRegistration extends Suite with TestRegistration with Informing wi
    * @param testName the name of one test to run.
    * @param args the <code>Args</code> for this run
    * @return a <code>Status</code> object that indicates when the test started by this method has completed, and whether or not it failed .
-   * @throws IllegalArgumentException if <code>testName</code> is defined but a test with that name does not exist on this <code>fixture.FunSuite</code>
-   * @throws NullArgumentException if <code>testName</code> or <code>args</code> is <code>null</code>.
+   * @throws IllegalArgumentException if <code>testName</code> is defined but a test with that name does not exist on this <code>fixture.PropSpec</code>
+   * @throws NullArgumentException if any of <code>testName</code> or <code>args</code> is <code>null</code>.
    */
   protected override def runTest(testName: String, args: Args): Status = {
 
@@ -207,8 +209,8 @@ trait FunSuiteRegistration extends Suite with TestRegistration with Informing wi
   }
 
   /**
-   * A <code>Map</code> whose keys are <code>String</code> tag names to which tests in this <code>fixture.FunSuite</code> belong, and values
-   * the <code>Set</code> of test names that belong to each tag. If this <code>fixture.FunSuite</code> contains no tags, this method returns an empty
+   * A <code>Map</code> whose keys are <code>String</code> tag names to which tests in this <code>fixture.PropSpec</code> belong, and values
+   * the <code>Set</code> of test names that belong to each tag. If this <code>fixture.PropSpec</code> contains no tags, this method returns an empty
    * <code>Map</code>.
    *
    * <p>
@@ -225,7 +227,7 @@ trait FunSuiteRegistration extends Suite with TestRegistration with Informing wi
 
   /**
    * <p>
-   * Run zero to many of this <code>fixture.FunSuiteLike</code>'s tests.
+   * Run zero to many of this <code>fixture.PropSpecLike</code>'s tests.
    * </p>
    *
    * <p>
@@ -267,18 +269,18 @@ trait FunSuiteRegistration extends Suite with TestRegistration with Informing wi
   }
 
   override def run(testName: Option[String], args: Args): Status = {
-    runImpl(thisSuite, testName, args: Args, super.run)
+    runImpl(thisSuite, testName, args, super.run)
   }
 
   /**
    * Registers shared tests.
    *
    * <p>
-   * This method enables the following syntax for shared tests in a <code>fixture.FunSuite</code>:
+   * This method enables the following syntax for shared tests in a <code>fixture.PropSpec</code>:
    * </p>
    *
    * <pre class="stHighlight">
-   * testsFor(nonEmptyStack(lastValuePushed))
+   * propertiesFor(nonEmptyStack(lastValuePushed))
    * </pre>
    *
    * <p>
@@ -286,11 +288,13 @@ trait FunSuiteRegistration extends Suite with TestRegistration with Informing wi
    * Because the parameter passed to it is
    * type <code>Unit</code>, the expression will be evaluated before being passed, which
    * is sufficient to register the shared tests. For examples of shared tests, see the
-   * <a href="../FunSuite.html#SharedTests">Shared tests section</a> in the main documentation for
-   * trait <code>FunSuite</code>.
+   * <a href="../PropSpec.html#SharedTests">Shared tests section</a> in the main documentation for
+   * trait <code>PropSpec</code>.
    * </p>
+   *
+   * @param unit a <code>Unit</code>
    */
-  protected def testsFor(unit: Unit) {}
+  protected def propertiesFor(unit: Unit) {}
 
   import scala.language.implicitConversions
 
@@ -325,9 +329,9 @@ trait FunSuiteRegistration extends Suite with TestRegistration with Informing wi
   /**
    * Suite style name.
    *
-   * @return <code>org.scalatest.fixture.FunSuite</code>
+   * @return <code>org.scalatest.fixture.PropSpec</code>
    */
-  final override val styleName: String = "org.scalatest.fixture.FunSuite"
+  final override val styleName: String = "org.scalatest.fixture.PropSpec"
 
   override def testDataFor(testName: String, theConfigMap: ConfigMap = ConfigMap.empty): TestData = createTestDataFor(testName, theConfigMap, this)
 }
