+++ scalatest/src/main/scala/org/scalatest/fixture/WordSpecLike.scala
@@ -16,4 +16,4 @@
 package org.scalatest.fixture
 
 //SCALATESTJS-ONLY @scala.scalajs.js.annotation.JSExportDescendentClasses(ignoreInvalidDescendants = true)
-trait WordSpecLike extends WordSpecRegistration with ClassicTests
+trait WordSpecLike extends WordSpecOf[Any]
